+++ +dat/expFilePath.m
@@ -20,11 +20,19 @@ function [fullpath, filename] = expFilePath(varargin)
 
 % 2013-03 CB created
 
-assert(length(varargin) > 1, 'Error: Not enough arguments supplied.')
+assert(length(varargin) > 1, 'Rigbox:dat:expFilePath:NotEnoughInputs',...
+  'Not enough input arguments.')
 
 parsed = catStructs(regexp(varargin{1}, dat.expRefRegExp, 'names'));
 if isempty(parsed) % Subject, not ref
-  if nargin > 4
+  if nargin < 3
+    error('Rigbox:dat:expFilePath:NotEnoughInputs', ...
+      ['Not enough input arguments; check expRef formatted correcly ' ...
+      'or enter subject, date and sequence as separate arguments'])
+  elseif nargin == 3
+    error('Rigbox:dat:expFilePath:NotEnoughInputs', ...
+      'Not enough input arguments; missing file type')
+  elseif nargin > 4
     location = varargin{5};
     varargin(5) = [];
   else
@@ -32,13 +40,16 @@ if isempty(parsed) % Subject, not ref
   end
   typeIdx = 4;
 else % Ref, not subject
-  typeIdx = 2;
-  if nargin > 2
+  if nargin < 2
+    error('Rigbox:dat:expFilePath:NotEnoughInputs', ...
+      'Not enough input arguments; missing file type')
+  elseif nargin > 2
     location = varargin{3};
     varargin(3) = [];
   else
     location = {};
   end
+  typeIdx = 2;
 end
 
 % tabulate the args to get complete rows