+++ tests/fixtures/util/+vis/init.m
@@ -1,0 +1,2 @@
+function occ = init(~)
+occ = struct;