+++ tests/fixtures/util/MockDialog.m
@@ -92,7 +92,7 @@ classdef MockDialog < handle
       
       % Check we're in test mode, throw warning if not
       if ~obj.InTest
-        warning('MockDialog:newCall:InTestFalse', ...
+        warning('Rigbox:MockDialog:newCall:notInTest', ...
           ['MockDialog method called whilst InTest flag set to false. ' ...
           'Check paths or set flag to true to avoid this message'])
       end
@@ -142,7 +142,7 @@ classdef MockDialog < handle
         answer = answer.first;
         obj.Dialogs(key) = obj.Dialogs(key).rest;
       elseif isa(answer, 'fun.EmptySeq')
-        warning('MockDialog:NewCall:EmptySeq', ...
+        warning('Rigbox:MockDialog:newCall:EmptySeq', ...
           'End of input sequence, using default input instead')
         answer = def;
       end
@@ -163,7 +163,8 @@ classdef MockDialog < handle
         key = [];
         return
       elseif ~obj.UseDefaults
-        assert(obj.Dialogs.Count > 0, 'MockDialog:newCall:NoValuesSet', ...
+        assert(obj.Dialogs.Count > 0, ...
+          'Rigbox:MockDialog:newCall:behaviourNotSet', ...
           'No values saved in Dialogs property')
       end
       key = obj.NumCalls;